requirements.txt
asgiref==3.7.2
Django==4.2.7
django-debug-toolbar==4.3.0
django-filter==23.5
djangorestframework==3.14.0
psycopg==3.1.17
psycopg-binary==3.1.17
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
